requirements.txt
PyYAML>=5
requests==2.20.1